update search func and button size
